refactor: extract drop indicator render
